requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
PyPDF2==3.0.1
pdfplumber==0.10.3
requests==2.31.0
werkzeug==3.0.1
supabase==2.0.3
python-dotenv
openai>=1.0.0
google-generativeai>=0.3.0 